pyproject.toml
[tool.poetry]
name = "growing-sparse-snn"
version = "0.1.0"
description = "Digital organism with dynamic growing spiking neural networks"
authors = ["nfriacowboy"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
norse = "^1.0.0"
numpy = "^1.24.0"
gymnasium = "^0.29.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"
pandas = "^2.1.0"
pyyaml = "^6.0"
prometheus-client = "^0.19.0"
tqdm = "^4.66.0"
tensorboard = "^2.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
ipython = "^8.18.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
